Refactor server and build logic for improved routing and error handling
- Moved routing and handler creation logic from `server.tsx` to `_server.tsx` for better organization.
- Introduced `createHandlers` and `buildApp` functions to streamline the creation of Hono applications with React Router integration.
- Enhanced error handling in server responses, including support for action and loader headers.
- Updated tests to cover new routing features, including index, parameterized, and catchall routes.
- Cleaned up unused imports and functions in `build.ts` and `server.tsx` for better maintainability.
