+++ build.ts
@@ -3,374 +3,21 @@ import {
   denoResolverPlugin,
 } from "@luca/esbuild-deno-loader";
 import { delay } from "@std/async/delay";
-import { sortBy } from "@std/collections/sort-by";
-import { ensureDir, exists, walk } from "@std/fs";
+import { ensureDir, exists } from "@std/fs";
 import * as path from "@std/path";
 import { toText } from "@std/streams";
 import * as esbuild from "esbuild";
 
 import { isProduction } from "@udibo/juniper/utils/env";
 
+import {
+  type GeneratedRoute,
+  generatedRouteObjectToString,
+  processClientDirectory,
+  processDirectory,
+} from "./_build.ts";
 import { deno } from "./deno.ts";
 import { startActiveSpan } from "./utils/_otel.ts";
-
-interface GeneratedRoute {
-  path: string;
-  main?: string;
-  index?: string;
-  catchall?: string;
-  children?: GeneratedRoute[];
-}
-
-interface DirRouteProperties {
-  main?: string;
-  index?: string;
-  catchall?: string;
-  fileModuleChildren: GeneratedRoute[];
-  parameterizedChildren: GeneratedRoute[];
-  directoryChildren: GeneratedRoute[];
-}
-
-function isMainRoute(fileName: string): boolean {
-  return fileName === "main.ts";
-}
-
-function isIndexRoute(fileName: string): boolean {
-  return fileName === "index.ts";
-}
-
-function isCatchallRoute(fileName: string): boolean {
-  return fileName === "[...].ts";
-}
-
-function isDynamicRoute(fileName: string): boolean {
-  return fileName.startsWith("[") && fileName.endsWith("].ts") &&
-    !isCatchallRoute(fileName);
-}
-
-function getDynamicRouteParam(fileName: string): string {
-  return fileName.slice(1, -4);
-}
-
-function isDynamicDirName(dirName: string): boolean {
-  return dirName.startsWith("[") && dirName.endsWith("]");
-}
-
-function getDynamicDirParam(dirName: string): string {
-  return dirName.slice(1, -1);
-}
-
-function isRegularFileModule(fileName: string): boolean {
-  return fileName.endsWith(".ts") &&
-    !isMainRoute(fileName) &&
-    !isIndexRoute(fileName) &&
-    !isCatchallRoute(fileName) &&
-    !isDynamicRoute(fileName) &&
-    !fileName.startsWith("_") &&
-    !fileName.includes(".test");
-}
-
-function isClientMainRoute(fileName: string): boolean {
-  return fileName === "main.tsx";
-}
-
-function isClientIndexRoute(fileName: string): boolean {
-  return fileName === "index.tsx";
-}
-
-function isClientCatchallRoute(fileName: string): boolean {
-  return fileName === "[...].tsx";
-}
-
-function isClientDynamicRoute(fileName: string): boolean {
-  return fileName.startsWith("[") && fileName.endsWith("].tsx") &&
-    !isClientCatchallRoute(fileName);
-}
-
-function getClientDynamicRouteParam(fileName: string): string {
-  return fileName.slice(1, -5);
-}
-
-function isRegularClientFileModule(fileName: string): boolean {
-  return fileName.endsWith(".tsx") &&
-    !isClientMainRoute(fileName) &&
-    !isClientIndexRoute(fileName) &&
-    !isClientCatchallRoute(fileName) &&
-    !isClientDynamicRoute(fileName) &&
-    !fileName.startsWith("_") &&
-    !fileName.includes(".test");
-}
-
-async function processDirectory(
-  absoluteDirPath: string,
-  importPathBase: string,
-): Promise<DirRouteProperties> {
-  const properties: DirRouteProperties = {
-    fileModuleChildren: [],
-    parameterizedChildren: [],
-    directoryChildren: [],
-  };
-
-  let dirEntries = [];
-  let firstEntrySkipped = false;
-  for await (
-    const entry of walk(absoluteDirPath, {
-      maxDepth: 1,
-      includeDirs: true,
-      includeFiles: true,
-      followSymlinks: false,
-    })
-  ) {
-    if (!firstEntrySkipped) {
-      firstEntrySkipped = true;
-      continue;
-    }
-    dirEntries.push(entry);
-  }
-
-  const entryPriority = (name: string): number => {
-    if (isMainRoute(name)) return 0;
-    if (isIndexRoute(name)) return 1;
-    if (isCatchallRoute(name)) return 2;
-    if (isDynamicRoute(name)) return 3;
-    return 4;
-  };
-
-  dirEntries = sortBy(
-    dirEntries,
-    (e) => `${e.isFile ? "0" : "1"}-${entryPriority(e.name)}-${e.name}`,
-  );
-
-  for (const entry of dirEntries) {
-    const currentFileImportPath = `${importPathBase}/${entry.name}`;
-    const moduleImport = `await import("${currentFileImportPath}")`;
-
-    if (entry.isFile && entry.name.endsWith(".ts")) {
-      if (isMainRoute(entry.name)) {
-        properties.main = moduleImport;
-      } else if (isIndexRoute(entry.name)) {
-        properties.index = moduleImport;
-      } else if (isCatchallRoute(entry.name)) {
-        properties.catchall = moduleImport;
-      } else if (isDynamicRoute(entry.name)) {
-        const paramName = getDynamicRouteParam(entry.name);
-        properties.parameterizedChildren.push({
-          path: `:${paramName}`,
-          main: moduleImport,
-        });
-      } else if (isRegularFileModule(entry.name)) {
-        const routeName = entry.name.slice(0, -3);
-        properties.fileModuleChildren.push({
-          path: routeName,
-          main: moduleImport,
-        });
-      }
-    } else if (entry.isDirectory) {
-      // Skip conventionally private directories like _components, _utils, etc.
-      if (entry.name.startsWith("_")) continue;
-
-      const subDirAbsolutePath = path.join(absoluteDirPath, entry.name);
-      const subDirImportPathBase = `${importPathBase}/${entry.name}`;
-      const subDirProps = await processDirectory(
-        subDirAbsolutePath,
-        subDirImportPathBase,
-      );
-
-      const children = [
-        ...subDirProps.fileModuleChildren,
-        ...subDirProps.directoryChildren,
-        ...subDirProps.parameterizedChildren,
-      ];
-
-      let routePathSegment = entry.name;
-      const isDynamicDir = isDynamicDirName(entry.name);
-      if (isDynamicDir) {
-        routePathSegment = `:${getDynamicDirParam(entry.name)}`;
-      }
-
-      const dirRoute: GeneratedRoute = {
-        path: routePathSegment,
-      };
-      if (subDirProps.main) dirRoute.main = subDirProps.main;
-      if (subDirProps.index) dirRoute.index = subDirProps.index;
-      if (subDirProps.catchall) dirRoute.catchall = subDirProps.catchall;
-      if (children.length > 0) dirRoute.children = children;
-
-      if (isDynamicDir) {
-        properties.parameterizedChildren.push(dirRoute);
-      } else {
-        properties.directoryChildren.push(dirRoute);
-      }
-    }
-  }
-  return properties;
-}
-
-/**
- * Processes a directory for client routes (.tsx files).
- * Similar to processDirectory but generates lazy loading syntax for client routes.
- */
-async function processClientDirectory(
-  absoluteDirPath: string,
-  importPathBase: string,
-  isRootLevel: boolean,
-): Promise<DirRouteProperties> {
-  const properties: DirRouteProperties = {
-    fileModuleChildren: [],
-    parameterizedChildren: [],
-    directoryChildren: [],
-  };
-
-  let dirEntries = [];
-  let firstEntrySkipped = false;
-  for await (
-    const entry of walk(absoluteDirPath, {
-      maxDepth: 1,
-      includeDirs: true,
-      includeFiles: true,
-      followSymlinks: false,
-    })
-  ) {
-    if (!firstEntrySkipped) {
-      firstEntrySkipped = true;
-      continue;
-    }
-    dirEntries.push(entry);
-  }
-
-  const entryPriority = (name: string): number => {
-    if (isClientMainRoute(name)) return 0;
-    if (isClientIndexRoute(name)) return 1;
-    if (isClientCatchallRoute(name)) return 2;
-    if (isClientDynamicRoute(name)) return 3;
-    return 4;
-  };
-
-  dirEntries = sortBy(
-    dirEntries,
-    (e) => `${e.isFile ? "0" : "1"}-${entryPriority(e.name)}-${e.name}`,
-  );
-
-  for (const entry of dirEntries) {
-    const currentFileImportPath = `${importPathBase}/${entry.name}`;
-    const lazyImport = `() => import("${currentFileImportPath}")`;
-    const directImport = `await import("${currentFileImportPath}")`;
-
-    if (entry.isFile && entry.name.endsWith(".tsx")) {
-      if (isClientMainRoute(entry.name)) {
-        // Only root-level main routes use await import, nested main routes use lazy import
-        properties.main = isRootLevel ? directImport : lazyImport;
-      } else if (isClientIndexRoute(entry.name)) {
-        properties.index = lazyImport; // Index routes use lazy import
-      } else if (isClientCatchallRoute(entry.name)) {
-        properties.catchall = lazyImport; // Catchall routes use lazy import
-      } else if (isClientDynamicRoute(entry.name)) {
-        const paramName = getClientDynamicRouteParam(entry.name);
-        properties.parameterizedChildren.push({
-          path: `:${paramName}`,
-          main: lazyImport, // Dynamic routes use lazy import
-        });
-      } else if (isRegularClientFileModule(entry.name)) {
-        const routeName = entry.name.slice(0, -4); // Remove .tsx
-        properties.fileModuleChildren.push({
-          path: routeName,
-          main: lazyImport, // Regular file modules use lazy import
-        });
-      }
-    } else if (entry.isDirectory) {
-      // Skip conventionally private directories like _components, _utils, etc.
-      if (entry.name.startsWith("_")) continue;
-
-      const subDirAbsolutePath = path.join(absoluteDirPath, entry.name);
-      const subDirImportPathBase = `${importPathBase}/${entry.name}`;
-      const subDirProps = await processClientDirectory(
-        subDirAbsolutePath,
-        subDirImportPathBase,
-        false,
-      );
-
-      const children = [
-        ...subDirProps.fileModuleChildren,
-        ...subDirProps.directoryChildren,
-        ...subDirProps.parameterizedChildren,
-      ];
-
-      // Only include directories that have .tsx files (main, index, catchall, or children)
-      const hasClientRoutes = subDirProps.main || subDirProps.index ||
-        subDirProps.catchall || children.length > 0;
-      if (!hasClientRoutes) continue;
-
-      let routePathSegment = entry.name;
-      const isDynamicDir = isDynamicDirName(entry.name);
-      if (isDynamicDir) {
-        routePathSegment = `:${getDynamicDirParam(entry.name)}`;
-      }
-
-      const dirRoute: GeneratedRoute = {
-        path: routePathSegment,
-      };
-      if (subDirProps.main) dirRoute.main = subDirProps.main;
-      if (subDirProps.index) dirRoute.index = subDirProps.index;
-      if (subDirProps.catchall) dirRoute.catchall = subDirProps.catchall;
-      if (children.length > 0) dirRoute.children = children;
-
-      if (isDynamicDir) {
-        properties.parameterizedChildren.push(dirRoute);
-      } else {
-        properties.directoryChildren.push(dirRoute);
-      }
-    }
-  }
-  return properties;
-}
-
-/**
- * Converts a generated route object to a string.
- *
- * @param obj The generated route object to convert to a string.
- * @param indentLevel The level of indentation to use when converting the object to a string.
- * @returns The generated route object as a string.
- */
-function generatedRouteObjectToString(
-  obj: GeneratedRoute | string | GeneratedRoute[],
-  indentLevel: number = 0,
-): string {
-  const indent = "  ".repeat(indentLevel);
-  const nextIndent = "  ".repeat(indentLevel + 1);
-
-  if (typeof obj === "string") {
-    if (obj.startsWith("await import(") || obj.startsWith("() => import(")) {
-      return obj;
-    }
-    return `"${
-      obj.replace(/\\/g, "\\\\").replace(/"/g, '\\"').replace(/\n/g, "\\n")
-    }"`;
-  }
-  if (Array.isArray(obj)) {
-    const arrElements = obj.map((item) =>
-      generatedRouteObjectToString(item, indentLevel + 1)
-    ).join(`,\n${nextIndent}`);
-    return `[\n${nextIndent}${arrElements}\n${indent}]`;
-  }
-  if (typeof obj === "object" && obj !== null) {
-    const props = Object.entries(obj)
-      .filter(([, value]) => {
-        return value !== undefined;
-      })
-      .map(([key, value]) =>
-        `${nextIndent}${key}: ${
-          generatedRouteObjectToString(value, indentLevel + 1)
-        }`
-      )
-      .join(",\n");
-
-    return `{\n${props}\n${indent}}`;
-  }
-  // Fallback for type checking, should be unreachable with correct input types
-  throw new Error(
-    `Unhandled type in generatedRouteObjectToString: ${typeof obj}`,
-  );
-}
 
 const fmtCommand = new Deno.Command(Deno.execPath(), {
   args: ["fmt", "-"],
